fixed bugs at all news controllers and fix deletepersonality

+++ controllers/news/deleteNews.js
@@ -1,19 +1,21 @@
 const { News } = require("../../models/news");
 const { deleteImgInCloudinary } = require("../../utils");
+const { HttpError } = require("../../helpers");
 
 const deleteNews = async (req, res) => {
   const { id: newsId } = req.params;
-  const { photos } = await News.findOne({ _id: newsId });
+  const news = await News.findOne({ _id: newsId });
 
-  for (const photo of photos) {
+  if (!news) {
+    throw HttpError(404, "Not Found");
+  }
+
+  for (const photo of news.photos) {
     deleteImgInCloudinary(photo);
   }
 
-  const result = await News.findByIdAndDelete(newsId);
+  await News.findByIdAndDelete(newsId);
 
-  if (!result) {
-    throw HttpError(404, "Not Found");
-  }
   res.status(204).json();
 };
